+++ src/pages/index.tsx
@@ -1,4 +1,5 @@
 // Related third party imports.
+import Head from "next/head";
 import { useContext, useEffect, useState } from "react";
 import {
   Upload,
@@ -16,20 +17,20 @@ import {
 } from "antd";
 import type { ColumnsType } from "antd/es/table";
 import { UploadOutlined, PlusOutlined } from "@ant-design/icons";
-import { getDownloadURL, list, ref } from "firebase/storage";
-import { getDocs, updateDoc, addDoc } from "firebase/firestore/lite";
+import { getDocs, addDoc, Timestamp } from "firebase/firestore/lite";
+import dayjs from "dayjs";
 
 // Local application/library specific imports.
+import styles from "./index.module.css";
 import type {
   BeefNoodleComment,
   BeefNoodleCommentFromFirestore,
 } from "../types";
 import GlobalContext from "../contexts/GlobalContext";
-import getCollection from "../utils/getCollection";
 import ImageDialogCarousel from "../components/ImageDialogCarousel";
-import Head from "next/head";
 
 // Stateless vars declare.
+const today = dayjs();
 const { Header, Content } = Layout;
 const { Title } = Typography;
 const { TextArea } = Input;
@@ -39,12 +40,18 @@ export default Home;
 function Home() {
   const [commentModalOpen, toggleCommentModalOpen] = useState(false);
   const [dataSource, setDataSource] = useState<BeefNoodleComment[]>([]);
-  const { firestore, firebaseStorage } = useContext(GlobalContext);
+  const { collectionRef } = useContext(GlobalContext);
+  const [beefCommentForm] = Form.useForm();
+  const [visitDate, setVisitDate] = useState(today);
+  const [addCommentLoading, toggleAddCommentLoading] = useState(false);
   const columns: ColumnsType<BeefNoodleComment> = [
     {
       title: "分數",
       dataIndex: "score",
       sorter: (a, b) => a.score - b.score,
+      render: (val: BeefNoodleComment["score"]) => (
+        <b className={styles.scoreTCell}>{val}</b>
+      ),
     },
     {
       title: "店名",
@@ -53,6 +60,10 @@ function Home() {
     {
       title: "造訪日期",
       dataIndex: "visitDate",
+      defaultSortOrder: "ascend",
+      sorter: (a, b) =>
+        Date.parse(a.visitDate.toISOString()) -
+        Date.parse(b.visitDate.toISOString()),
       render: (val: BeefNoodleComment["visitDate"]) => val.toLocaleDateString(),
     },
     {
@@ -75,6 +86,9 @@ function Home() {
       title: "麵條分數",
       dataIndex: "noodleScore",
       sorter: (a, b) => (a.noodleScore || 0) - (b.noodleScore || 0),
+      render: (val: BeefNoodleComment["noodleScore"]) => (
+        <b className={styles.scoreTCell}>{val}</b>
+      ),
     },
     {
       title: "麵條描述",
@@ -84,33 +98,21 @@ function Home() {
       title: "牛肉分數",
       dataIndex: "beefScore",
       sorter: (a, b) => (a.beefScore || 0) - (b.beefScore || 0),
+      render: (val: BeefNoodleComment["beefScore"]) => (
+        <b className={styles.scoreTCell}>{val}</b>
+      ),
     },
     {
       title: "牛肉描述",
       dataIndex: "beefDescription",
     },
     {
-      title: "牛筋分數",
-      dataIndex: "tendonScore",
-      sorter: (a, b) => (a.tendonScore || 0) - (b.tendonScore || 0),
-    },
-    {
-      title: "牛筋描述",
-      dataIndex: "tendonDescription",
-    },
-    {
-      title: "牛肚分數",
-      dataIndex: "tripeScore",
-      sorter: (a, b) => (a.tripeScore || 0) - (b.tripeScore || 0),
-    },
-    {
-      title: "牛肚描述",
-      dataIndex: "tripeDescription",
-    },
-    {
       title: "湯頭分數",
       dataIndex: "soupScore",
       sorter: (a, b) => (a.soupScore || 0) - (b.soupScore || 0),
+      render: (val: BeefNoodleComment["soupScore"]) => (
+        <b className={styles.scoreTCell}>{val}</b>
+      ),
     },
     {
       title: "湯頭描述",
@@ -121,18 +123,89 @@ function Home() {
       dataIndex: "overallDescription",
     },
     {
-      title: "是否願意再次造訪",
+      title: "再次造訪",
       dataIndex: "wantToVisitAgain",
-      render: (val: BeefNoodleComment["wantToVisitAgain"]) =>
-        val ? "是" : "否",
+      render: (val: BeefNoodleComment["wantToVisitAgain"]) => (
+        <b className={styles.booleanTCell}>{val ? "是" : "否"}</b>
+      ),
       sorter: (a, b) => (a.wantToVisitAgain ? 1 : -1),
     },
+    {
+      title: "牛筋分數",
+      dataIndex: "tendonScore",
+      sorter: (a, b) => (a.tendonScore || 0) - (b.tendonScore || 0),
+      render: (val: BeefNoodleComment["tendonScore"]) => (
+        <b className={styles.scoreTCell}>{val}</b>
+      ),
+    },
+    {
+      title: "牛筋描述",
+      dataIndex: "tendonDescription",
+    },
+    {
+      title: "牛肚分數",
+      dataIndex: "tripeScore",
+      sorter: (a, b) => (a.tripeScore || 0) - (b.tripeScore || 0),
+      render: (val: BeefNoodleComment["tripeScore"]) => (
+        <b className={styles.scoreTCell}>{val}</b>
+      ),
+    },
+    {
+      title: "牛肚描述",
+      dataIndex: "tripeDescription",
+    },
   ];
+  /**
+   * @todo 照片上傳還沒弄
+   */
+  async function insertOneComment() {
+    const { getFieldValue } = beefCommentForm;
+    const comment: BeefNoodleCommentFromFirestore = {
+      score: parseInt(getFieldValue("score")),
+      storeName: String(getFieldValue("storeName")),
+      visitDate: Timestamp.fromDate(visitDate.toDate()),
+      itemName: String(getFieldValue("itemName")),
+      itemPrice: parseInt(getFieldValue("itemPrice")),
+      images: [],
+      wantToVisitAgain: Boolean(getFieldValue("wantToVisitAgain")),
+      noodleScore: getFieldValue("noodleScore"),
+      noodleDescription: getFieldValue("noodleDescription"),
+      beefScore: getFieldValue("beefScore"),
+      beefDescription: getFieldValue("beefDescription"),
+      tendonScore: getFieldValue("tendonScore"),
+      tendonDescription: getFieldValue("tendonDescription"),
+      tripeScore: getFieldValue("tripeScore"),
+      tripeDescription: getFieldValue("tripeDescription"),
+      soupScore: getFieldValue("soupScore"),
+      soupDescription: getFieldValue("soupDescription"),
+      overallDescription: getFieldValue("overallDescription"),
+    };
+    if (!comment.noodleScore) delete comment.noodleScore;
+    if (!comment.noodleDescription) delete comment.noodleDescription;
+    if (!comment.beefScore) delete comment.beefScore;
+    if (!comment.beefDescription) delete comment.beefDescription;
+    if (!comment.tendonScore) delete comment.tendonScore;
+    if (!comment.tendonDescription) delete comment.tendonDescription;
+    if (!comment.tripeScore) delete comment.tripeScore;
+    if (!comment.tripeDescription) delete comment.tripeDescription;
+    if (!comment.soupScore) delete comment.soupScore;
+    if (!comment.soupDescription) delete comment.soupDescription;
+    if (!comment.overallDescription) delete comment.overallDescription;
+    toggleAddCommentLoading(true);
+    await addDoc(collectionRef, comment);
+    setDataSource((prev) => [
+      ...prev,
+      {
+        ...comment,
+        key: prev.length,
+        visitDate: new Date(comment.visitDate.seconds * 1000),
+      },
+    ]);
+    beefCommentForm.resetFields();
+    toggleAddCommentLoading(false);
+    toggleCommentModalOpen(false);
+  }
   useEffect(function getAllComments() {
-    const collectionRef = getCollection<BeefNoodleCommentFromFirestore>(
-      firestore,
-      "beefNoodleComments"
-    );
     getDocs(collectionRef)
       .then((querySnapshot) => {
         const beefNoodleComments: BeefNoodleComment[] = querySnapshot.docs.map(
@@ -141,7 +214,7 @@ function Home() {
             return {
               ...document,
               key: idx,
-              visitDate: new Date(document.visitDate.seconds),
+              visitDate: new Date(document.visitDate.seconds * 1000),
             };
           }
         );
@@ -196,77 +269,111 @@ function Home() {
       <Content>
         <Table
           bordered
-          style={{ wordBreak: "keep-all" }}
-          scroll={{ x: true }}
+          className={styles.table}
+          // header 64px + pagination 64px + thead 55px + 1px
+          scroll={{ x: true, y: "calc(100vh - 64px - 64px - 55px - 1px)" }}
           dataSource={dataSource}
           columns={columns}
-          pagination={false}
         ></Table>
         <Modal
           title="新增評論"
           open={commentModalOpen}
-          onOk={() => toggleCommentModalOpen(false)}
+          footer={null}
+          cancelButtonProps={{ disabled: addCommentLoading }}
           onCancel={() => toggleCommentModalOpen(false)}
         >
-          <Form labelCol={{ span: 4 }} wrapperCol={{ span: 20 }}>
-            <Form.Item label="店名：">
-              <Input></Input>
+          <Form
+            form={beefCommentForm}
+            labelCol={{ span: 4 }}
+            wrapperCol={{ span: 20 }}
+            className={styles.beefCommentForm}
+          >
+            <Form.Item required label="店名：" name="storeName">
+              <Input required></Input>
             </Form.Item>
-            <Form.Item label="分數：">
+            <Form.Item required label="分數：" name="score">
+              <InputNumber
+                required
+                controls={false}
+                min={0}
+                max={100}
+              ></InputNumber>
+            </Form.Item>
+            <Form.Item required label="造訪日期：">
+              <DatePicker
+                aria-required
+                allowClear={false}
+                value={visitDate}
+                onChange={(day) => day && setVisitDate(day)}
+              ></DatePicker>
+            </Form.Item>
+            <Form.Item required label="品項：" name="itemName">
+              <Input required></Input>
+            </Form.Item>
+            <Form.Item required label="價格：" name="itemPrice">
+              <InputNumber required controls={false}></InputNumber>
+            </Form.Item>
+            <Form.Item required label="再度造訪：" name="wantToVisitAgain">
+              <Radio.Group>
+                <Radio value={1}>是</Radio>
+                <Radio value={0}>否</Radio>
+              </Radio.Group>
+            </Form.Item>
+            <Form.Item label="麵條分數：" name="noodleScore">
               <InputNumber controls={false} min={0} max={100}></InputNumber>
             </Form.Item>
-            <Form.Item label="造訪日期：">
-              <DatePicker></DatePicker>
-            </Form.Item>
-            <Form.Item label="品項：">
-              <Input></Input>
-            </Form.Item>
-            <Form.Item label="價格：">
-              <InputNumber controls={false}></InputNumber>
-            </Form.Item>
-            <Form.Item label="麵條分數：">
-              <InputNumber controls={false} min={0} max={100}></InputNumber>
-            </Form.Item>
-            <Form.Item label=" 麵條評論：">
+            <Form.Item label="麵條評論：" name="noodleDescription">
               <TextArea></TextArea>
             </Form.Item>
-            <Form.Item label="湯頭分數：">
+            <Form.Item label="湯頭分數：" name="soupScore">
               <InputNumber controls={false} min={0} max={100}></InputNumber>
             </Form.Item>
-            <Form.Item label="湯頭評論：">
+            <Form.Item label="湯頭評論：" name="soupDescription">
               <TextArea></TextArea>
             </Form.Item>
-            <Form.Item label="牛肉分數：">
+            <Form.Item label="牛肉分數：" name="beefScore">
               <InputNumber controls={false} min={0} max={100}></InputNumber>
             </Form.Item>
-            <Form.Item label="牛肉評論：">
+            <Form.Item label="牛肉評論：" name="beefDescription">
               <TextArea></TextArea>
             </Form.Item>
-            <Form.Item label="牛筋分數：">
+            <Form.Item label="牛筋分數：" name="tendonScore">
               <InputNumber controls={false} min={0} max={100}></InputNumber>
             </Form.Item>
-            <Form.Item label="牛筋評論：">
+            <Form.Item label="牛筋評論：" name="tendonDescription">
               <TextArea></TextArea>
             </Form.Item>
-            <Form.Item label="牛肚分數：">
+            <Form.Item label="牛肚分數：" name="tripeScore">
               <InputNumber controls={false} min={0} max={100}></InputNumber>
             </Form.Item>
-            <Form.Item label="牛肚評論：">
+            <Form.Item label="牛肚評論：" name="tripeDescription">
               <TextArea></TextArea>
             </Form.Item>
             <Form.Item label="圖片：">
-              <Upload>
+              <Upload onChange={(info) => console.log(info)}>
                 <Button icon={<UploadOutlined />}>上傳圖片</Button>
               </Upload>
             </Form.Item>
-            <Form.Item label="整體評論：">
+            <Form.Item label="整體評論：" name="overallDescription">
               <TextArea></TextArea>
             </Form.Item>
-            <Form.Item label="再度造訪：">
-              <Radio.Group>
-                <Radio>是</Radio>
-                <Radio>否</Radio>
-              </Radio.Group>
+            <Form.Item wrapperCol={{ span: 24 }} style={{ textAlign: "right" }}>
+              <Button
+                htmlType="button"
+                onClick={() => toggleCommentModalOpen(false)}
+                style={{ marginRight: "8px" }}
+                disabled={addCommentLoading}
+              >
+                取消
+              </Button>
+              <Button
+                htmlType="submit"
+                type="primary"
+                onClick={insertOneComment}
+                loading={addCommentLoading}
+              >
+                送出
+              </Button>
             </Form.Item>
           </Form>
         </Modal>
@@ -274,6 +381,7 @@ function Home() {
           icon={<PlusOutlined />}
           tooltip="新增評論"
           type="primary"
+          style={{ left: "24px", bottom: "16px" }}
           onClick={() => toggleCommentModalOpen(true)}
         ></FloatButton>
       </Content>